Added functionality of storing & displaying images
This functionality is meant so that users can attatch receipts to the restaurants, this can also be used as a receipt storing app.
Users can select an image from their local storage and input it, which then gets sent to the backend where multer is used to store
the image there and in the database as binary data. Then it is retrieved when it has to be displayed and transformed into base64.

Added frontend features of clicking on each block to display more information, and a basic google map to be developed for later.